+++ src/main/java/cn/nukkit/network/protocol/InventoryContentPacket.java
@@ -11,7 +11,7 @@ public class InventoryContentPacket extends DataPacket {
     public static final byte NETWORK_ID = ProtocolInfo.INVENTORY_CONTENT_PACKET;
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
