+++ src/main/java/cn/nukkit/network/protocol/UpdateBlockSyncedPacket.java
@@ -12,7 +12,7 @@ public class UpdateBlockSyncedPacket extends DataPacket {
 
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
